On failure, save a dummy file from the background page thumb capture module.
If a background page thumbnail capture fails (e.g. due to too heavy scripting), we should
write -something- to the thumbnail cache, because otherwise it will try again and again,
which is problematic for bad trap pages, that even if the user has left the page never
to return again, the thumbnail service may still try to capture, and fail.

This resolves the only problem for us in %592.