Revert "Treat *.jnlp as an executable class file, like *.jar"
Rationale: This was a Mozilla oversight and/or error.
This change has caused harm and is causing users to switch back to
Chrome, Safari or Edge for their WebStart needs.

JNLP is not an executable and should not be treated as such. JNLP should
be treated the same as any (e.g. Word) document and allowed to be opened
with the designated program.
A JNLP file will not cause execution on a system unless it has a valid
signature, and the user explicitly authorizes the launching based on
information provided by the signature. Moreover, there will even be a
check by the Java environment to see if the Java runtime (if there is
one) is current, and prompt the user to update if required.

This reverts commit 21495c58976e3cbbfe54d2e54d1fd67e36dff2a6 and
modifies ApplicationReputation.cpp to keep the list in sync (was a
discrepancy before).