MoonchildProductions%1251 - Part 15: fdlibm should provide definition for u_int32_t and u_int64_t.
https://bugzilla.mozilla.org/show_bug.cgi?id=1350355

u_int32_t is not an stdint.h type. Windows already requires this, Solaris needs it too. If someone has a nit with this approach, the alternatives include:

1. Just replacing every instance of u_int32_t with uint32_t.

2. Including <sys/types.h> for Solaris only, which does define this.

3. Changing the original ifdef to be WIN32 || XP_SOLARIS

But it really doesn't matter how you solve this problem, all of the approaches are functionally equivalent, and this one has been used in Firefox since version 55. As far as I can tell, all it does is apply a fix that was being done for Windows already to any platform that needs it.