Remove MOZ_MAIL_LEGACY flag except from ClassicTB2.
I think I got all the issues sorted with this, the extraneous endif and the misreading of one ifndef as an ifdef. I also went ahead and removed the flag from mail/themes/moz.build, so that all remaining references to MOZ_MAIL_LEGACY are now in the ClassicTB2 directory.